move damageutils to utils.entity
