+++ src/main/java/meteordevelopment/meteorclient/utils/entity/DamageUtils.java
@@ -3,7 +3,7 @@
  * Copyright (c) Meteor Development.
  */
 
-package meteordevelopment.meteorclient.utils.player;
+package meteordevelopment.meteorclient.utils.entity;
 
 import meteordevelopment.meteorclient.mixininterface.IVec3d;
 import meteordevelopment.meteorclient.utils.entity.EntityAttributeManager;
